# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RelaySMS Publisher",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/publications": {
      "post": {
        "tags": [
          "Publishing"
        ],
        "summary": "Publish",
        "description": "Queues an SMS payload for publishing.\n\nMulti-part payloads are kept until every segment arrives, then published. Payloads\nare tagged with protocol `https`; see Offline Publishing in the README for\n`OFFLINE_PUBLISH_ALLOWED_PROTOCOLS` and `tag`. Decryption and adapter errors\nhappen later in the queue, so they're logged, not returned.",
        "operationId": "create_publications_v1_publications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRestContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "The payload isn't base64, or isn't a RelaySMS payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/twilio-sms": {
      "post": {
        "tags": [
          "Publishing"
        ],
        "summary": "Twilio SMS webhook",
        "description": "Checks the Twilio signature, then queues the SMS for publishing.\n\nPayloads are tagged with protocol `sms`. Each SMS is also forwarded to\n`TWILIO_FORWARD_URLS_RAW` as Twilio's form and to `TWILIO_FORWARD_URLS_JSON` as\n`{\"sender\", \"text\", \"received_at\"}`.",
        "operationId": "twilio_incoming_sms_v1_twilio_sms_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "From": {
                    "type": "string",
                    "description": "Sender phone number"
                  },
                  "Body": {
                    "type": "string",
                    "description": "Base64-encoded SMS payload"
                  }
                },
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object",
                "required": [
                  "From",
                  "Body"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Empty TwiML, so Twilio sends no reply.",
            "content": {
              "text/xml": {
                "example": "<?xml version=\"1.0\" encoding=\"UTF-8\"?><Response />"
              }
            }
          },
          "400": {
            "description": "No From or Body field. The payload isn't base64, or isn't a RelaySMS payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "No valid X-Twilio-Signature.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "TWILIO_SMS_TRANSPORT_ENABLED isn't true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platforms": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "List platforms",
        "description": "Platforms users can link: those with an installed, enabled adapter.",
        "operationId": "get_platforms_v1_platforms_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50,
                  "pattern": "^[a-zA-Z0-9_-]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform name",
              "title": "Name"
            },
            "description": "Filter by platform name"
          },
          {
            "name": "proto_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by protocol ID",
              "title": "Proto Id"
            },
            "description": "Filter by protocol ID"
          },
          {
            "name": "cat_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category ID",
              "title": "Cat Id"
            },
            "description": "Filter by category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformManifest"
                  },
                  "title": "Response Get Platforms V1 Platforms Get"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platforms/{platform_name}/oauth/client-metadata.json": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "OAuth client metadata",
        "description": "Only for platforms with dynamic client registration, such as Bluesky.",
        "operationId": "get_platform_oauth_client_metadata_v1_platforms__platform_name__oauth_client_metadata_json_get",
        "parameters": [
          {
            "name": "platform_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]+$",
              "description": "Platform name",
              "title": "Platform Name"
            },
            "description": "Platform name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthClientMetadata"
                }
              }
            }
          },
          "404": {
            "description": "No enabled adapter for the platform, or the platform has no client metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/platforms/{platform_name}/oauth/callback": {
      "get": {
        "tags": [
          "Platforms"
        ],
        "summary": "OAuth callback",
        "description": "Shows the callback query parameters, as a redirect target for OAuth2 testing.",
        "operationId": "oauth_callback_v1_platforms__platform_name__oauth_callback_get",
        "parameters": [
          {
            "name": "platform_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]+$",
              "description": "Platform name",
              "title": "Platform Name"
            },
            "description": "Platform name"
          }
        ],
        "responses": {
          "200": {
            "description": "The query parameters as an HTML table.",
            "content": {
              "text/html": {}
            }
          },
          "404": {
            "description": "No enabled adapter for the platform, or the platform has no client metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "What went wrong, safe to show to users."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error response."
      },
...
      "OAuthClientMetadata": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "dpop_bound_access_tokens": {
            "type": "boolean",
            "title": "Dpop Bound Access Tokens"
          },
          "application_type": {
            "type": "string",
            "title": "Application Type"
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris"
          },
          "grant_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Grant Types"
          },
          "response_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Response Types"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "token_endpoint_auth_method": {
            "type": "string",
            "title": "Token Endpoint Auth Method"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "client_uri": {
            "type": "string",
            "title": "Client Uri"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "dpop_bound_access_tokens",
          "application_type",
          "redirect_uris",
          "grant_types",
          "response_types",
          "scope",
          "token_endpoint_auth_method",
          "client_name",
          "client_uri"
        ],
        "title": "OAuthClientMetadata"
      },
...
      "PlatformManifest": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "proto_id": {
            "type": "integer",
            "title": "Proto Id",
            "description": "0 = oauth2, 1 = pnba"
          },
          "cat_id": {
            "type": "integer",
            "title": "Cat Id",
            "description": "0 = email, 1 = message, 2 = text, 3 = bridge"
          },
          "auth_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Provider"
          },
          "supports_offline_first": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Offline First"
          },
          "icon_svg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Svg"
          },
          "icon_png": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Png"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "name",
          "proto_id",
          "cat_id"
        ],
        "title": "PlatformManifest"
      },
...
      "PublishContentResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Always null here; errors come back with a 4xx status."
          }
        },
        "type": "object",
        "title": "PublishContentResponse"
      },
      "PublishRestContentRequest": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address",
            "description": "Sender phone number in E.164 format",
            "examples": [
              "+12025550123"
            ]
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Base64-encoded SMS payload"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag",
            "description": "Shared secret for offline payloads, if the server sets one."
          }
        },
        "type": "object",
        "required": [
          "address",
          "text"
        ],
        "title": "PublishRestContentRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "SessionCookie": {
        "type": "apiKey",
        "description": "Set by POST /v1/auth/login.",
        "in": "cookie",
        "name": "relaysms_session"
      }
    }
  }
}